+++ internal/gophermart/orderprocessor/iface.go
@@ -2,5 +2,5 @@ package orderprocessor
 
 type processor interface {
 	GetUnfinishedOrders() (map[string]string, error)
-	UpdateOrderStatus(order, status string, accrual int) error
+	UpdateOrderStatus(order, status string, accrual float64) error
 }
